Transcribe-Translate-Recordings/requirements.txt
openai==0.27.0
pydub==0.25.1
python-dotenv==1.0.0
